Implement ThresholdFinder class
There were some major changes to the Lattice class, as well as Node.
Now they operate on node ids, not pointers. The reason is simple:
as we need to store an array of visited nodes in order to perform
DFS, a solution based on sets and pointers may (probably) be slightly
slower. However, I did not test this hypothesis properly and will
try out another possible implementation in the future.